requirements.txt
boto3==1.26.33
email-validator==1.3.0
fastapi==0.89.1
passlib==1.7.4
# psycopg2==2.9.5
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
python-multipart==0.0.5
shortuuid==1.0.11
SQLAlchemy==1.4.45
uvicorn==0.20.0
alembic==1.9.1
faker==16.3.0
# pytest-asyncio==0.20.3
pytest==7.2.1 
httpx==0.23.1
SQLAlchemy-Utils==0.39.0 
fastapi-pagination==0.11.2
pytz==2022.7.1
typeguard==2.13.3
